examples/python/requirements.txt
vaea-flash
solders>=0.20.0
python-dotenv>=1.0.0
